feat(tests): apt-283 add getAmountInt getAmountOut functions
