+++ src/constants/index.ts
@@ -17,3 +17,11 @@ export const NETWORKS_MODULES = {
   CoinInfo: `${COIN_INFO}`,
   CoinStore: `${COIN_STORE}`,
 };
+
+export const TOKENS_MAPPING: Record<string, string> = {
+  APTOS: '0x1::aptos_coin::AptosCoin',
+  USDT: '0xf22bede237a07e121b56d91a491eb7bcdfd1f5907926a9e58338f964a01b17fa::asset::USDT', //layerzero USDT
+  BTC: '0xae478ff7d83ed072dbc5e264250e67ef58f57c99d89b447efd8a0a2e8b2be76e::coin::T', // wormhole wrapped BTC
+  WETH: '0xcc8a89c8dce9693d354449f1f73e60e14e347417854f029db5bc8e7454008abb::coin::T', // wormhole WETH
+  USDC: '0xf22bede237a07e121b56d91a491eb7bcdfd1f5907926a9e58338f964a01b17fa::asset::USDC', // layerzero USDC
+};
